backend/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
pydantic
elasticsearch
celery
redis
python-dotenv
pandas